lab_kafka/requirements.txt
appnope==0.1.3
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kafka-python==2.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
urllib3==2.1.0
wcwidth==0.2.12
